refactor: move getSub to auth service
