+++ schema.graphql
@@ -458,6 +458,16 @@ type UpdateServerConfigPayload {
 }
 
 type Query {
+  authCheck: Boolean!
+  serverInvite(token: String!): ServerInvite!
+  serverInvites: [ServerInvite!]!
+  me: User!
+  user(id: Int, name: String): User!
+  users: [User!]!
+  usersByIds(ids: [Int!]!): [User!]!
+  isFirstUser: Boolean!
+  serverRole(id: Int!): ServerRole!
+  serverRoles: [ServerRole!]!
   post(id: Int!): Post!
   proposal(id: Int!): Proposal!
   group(id: Int, name: String): Group!
@@ -466,19 +476,9 @@ type Query {
   publicGroupsFeed: [FeedItem!]!
   groupRole(id: Int!): GroupRole!
   groupMemberRequest(groupId: Int!): GroupMemberRequest
-  me: User!
-  user(id: Int, name: String): User!
-  users: [User!]!
-  usersByIds(ids: [Int!]!): [User!]!
-  isFirstUser: Boolean!
-  serverRole(id: Int!): ServerRole!
-  serverRoles: [ServerRole!]!
   event(id: Int): Event!
   events(input: EventsInput!): [Event!]!
   groupRoles: [GroupRole!]!
-  authCheck: Boolean!
-  serverInvite(token: String!): ServerInvite!
-  serverInvites: [ServerInvite!]!
   canary: Canary!
   publicCanary: Canary
   serverConfig: ServerConfig!
@@ -490,16 +490,31 @@ input EventsInput {
 }
 
 type Mutation {
-  createComment(commentData: CreateCommentInput!): CreateCommentPayload!
-  updateComment(commentData: UpdateCommentInput!): UpdateCommentPayload!
-  deleteComment(id: Int!): Boolean!
+  login(input: LoginInput!): Boolean!
+  signUp(input: SignUpInput!): Boolean!
+  logOut: Boolean!
+  createServerInvite(serverInviteData: CreateServerInviteInput!): CreateServerInvitePayload!
+  deleteServerInvite(id: Int!): Boolean!
+  updateUser(userData: UpdateUserInput!): UpdateUserPayload!
+  deleteUser(id: Int!): Boolean!
+  followUser(id: Int!): FollowUserPayload!
+  unfollowUser(id: Int!): Boolean!
+  createServerRole(serverRoleData: CreateServerRoleInput!): CreateServerRolePayload!
+  updateServerRole(serverRoleData: UpdateServerRoleInput!): UpdateServerRolePayload!
+  deleteServerRole(id: Int!): Boolean!
+  deleteServerRoleMember(serverRoleMemberData: DeleteServerRoleMemberInput!): DeleteServerRoleMemberPayload!
+  deleteImage(id: Int!): Boolean!
   createPost(postData: CreatePostInput!): CreatePostPayload!
   updatePost(postData: UpdatePostInput!): UpdatePostPayload!
   deletePost(id: Int!): Boolean!
+  createComment(commentData: CreateCommentInput!): CreateCommentPayload!
+  updateComment(commentData: UpdateCommentInput!): UpdateCommentPayload!
+  deleteComment(id: Int!): Boolean!
+  createLike(likeData: CreateLikeInput!): CreateLikePayload!
+  deleteLike(likeData: DeleteLikeInput!): Boolean!
   createProposal(proposalData: CreateProposalInput!): CreateProposalPayload!
   updateProposal(proposalData: UpdateProposalInput!): UpdateProposalPayload!
   deleteProposal(id: Int!): Boolean!
-  deleteImage(id: Int!): Boolean!
   createVote(voteData: CreateVoteInput!): CreateVotePayload!
   updateVote(voteData: UpdateVoteInput!): UpdateVotePayload!
   deleteVote(id: Int!): Boolean!
@@ -512,16 +527,6 @@ type Mutation {
   approveGroupMemberRequest(id: Int!): ApproveGroupMemberRequestPayload!
   cancelGroupMemberRequest(id: Int!): Boolean!
   denyGroupMemberRequest(id: Int!): Boolean!
-  createLike(likeData: CreateLikeInput!): CreateLikePayload!
-  deleteLike(likeData: DeleteLikeInput!): Boolean!
-  updateUser(userData: UpdateUserInput!): UpdateUserPayload!
-  deleteUser(id: Int!): Boolean!
-  followUser(id: Int!): FollowUserPayload!
-  unfollowUser(id: Int!): Boolean!
-  createServerRole(serverRoleData: CreateServerRoleInput!): CreateServerRolePayload!
-  updateServerRole(serverRoleData: UpdateServerRoleInput!): UpdateServerRolePayload!
-  deleteServerRole(id: Int!): Boolean!
-  deleteServerRoleMember(serverRoleMemberData: DeleteServerRoleMemberInput!): DeleteServerRoleMemberPayload!
   createEvent(eventData: CreateEventInput!): CreateEventPayload!
   updateEvent(eventData: UpdateEventInput!): UpdateEventPayload!
   deleteEvent(id: Int!): Boolean!
@@ -532,28 +537,66 @@ type Mutation {
   updateGroupRole(groupRoleData: UpdateGroupRoleInput!): UpdateGroupRolePayload!
   deleteGroupRole(id: Int!): Boolean!
   deleteGroupRoleMember(groupRoleMemberData: DeleteGroupRoleMemberInput!): DeleteGroupRoleMemberPayload!
-  login(input: LoginInput!): Boolean!
-  signUp(input: SignUpInput!): Boolean!
-  logOut: Boolean!
-  createServerInvite(serverInviteData: CreateServerInviteInput!): CreateServerInvitePayload!
-  deleteServerInvite(id: Int!): Boolean!
   updateServerConfig(serverConfigData: UpdateServerConfigInput!): UpdateServerConfigPayload!
 }
 
-input CreateCommentInput {
-  postId: Int
-  proposalId: Int
-  body: String
-  images: [Upload!]
+input LoginInput {
+  email: String!
+  password: String!
+}
+
+input SignUpInput {
+  email: String!
+  name: String!
+  password: String!
+  confirmPassword: String!
+  profilePicture: Upload
+  inviteToken: String
 }
 
 """The `Upload` scalar type represents a file upload."""
 scalar Upload
 
-input UpdateCommentInput {
+input CreateServerInviteInput {
+  maxUses: Int
+  expiresAt: DateTime
+}
+
+input UpdateUserInput {
   id: Int!
-  body: String
-  images: [Upload!]
+  name: String!
+  bio: String!
+  profilePicture: Upload
+  coverPhoto: Upload
+}
+
+input CreateServerRoleInput {
+  name: String!
+  color: String!
+}
+
+input UpdateServerRoleInput {
+  id: Int!
+  name: String
+  color: String
+  selectedUserIds: [Int!]
+  permissions: ServerRolePermissionInput
+}
+
+input ServerRolePermissionInput {
+  manageRoles: Boolean
+  managePosts: Boolean
+  manageComments: Boolean
+  manageEvents: Boolean
+  manageSettings: Boolean
+  manageInvites: Boolean
+  createInvites: Boolean
+  removeMembers: Boolean
+}
+
+input DeleteServerRoleMemberInput {
+  serverRoleId: Int!
+  userId: Int!
 }
 
 input CreatePostInput {
@@ -567,6 +610,27 @@ input UpdatePostInput {
   id: Int!
   body: String
   images: [Upload!]
+}
+
+input CreateCommentInput {
+  postId: Int
+  proposalId: Int
+  body: String
+  images: [Upload!]
+}
+
+input UpdateCommentInput {
+  id: Int!
+  body: String
+  images: [Upload!]
+}
+
+input CreateLikeInput {
+  postId: Int
+}
+
+input DeleteLikeInput {
+  postId: Int
 }
 
 input CreateProposalInput {
@@ -658,51 +722,6 @@ input UpdateGroupConfigInput {
   privacy: String
 }
 
-input CreateLikeInput {
-  postId: Int
-}
-
-input DeleteLikeInput {
-  postId: Int
-}
-
-input UpdateUserInput {
-  id: Int!
-  name: String!
-  bio: String!
-  profilePicture: Upload
-  coverPhoto: Upload
-}
-
-input CreateServerRoleInput {
-  name: String!
-  color: String!
-}
-
-input UpdateServerRoleInput {
-  id: Int!
-  name: String
-  color: String
-  selectedUserIds: [Int!]
-  permissions: ServerRolePermissionInput
-}
-
-input ServerRolePermissionInput {
-  manageRoles: Boolean
-  managePosts: Boolean
-  manageComments: Boolean
-  manageEvents: Boolean
-  manageSettings: Boolean
-  manageInvites: Boolean
-  createInvites: Boolean
-  removeMembers: Boolean
-}
-
-input DeleteServerRoleMemberInput {
-  serverRoleId: Int!
-  userId: Int!
-}
-
 input CreateEventInput {
   name: String!
   description: String!
@@ -756,25 +775,6 @@ input UpdateGroupRoleInput {
 input DeleteGroupRoleMemberInput {
   groupRoleId: Int!
   userId: Int!
-}
-
-input LoginInput {
-  email: String!
-  password: String!
-}
-
-input SignUpInput {
-  email: String!
-  name: String!
-  password: String!
-  confirmPassword: String!
-  profilePicture: Upload
-  inviteToken: String
-}
-
-input CreateServerInviteInput {
-  maxUses: Int
-  expiresAt: DateTime
 }
 
 input UpdateServerConfigInput {
