+++ src/context/context.module.ts
@@ -1,4 +1,5 @@
 import { Module } from '@nestjs/common';
+import { AuthModule } from '../auth/auth.module';
 import { CommentsModule } from '../comments/comments.module';
 import { DataloaderModule } from '../dataloader/dataloader.module';
 import { EventsModule } from '../events/events.module';
@@ -17,6 +18,7 @@ import { ContextService } from './context.service';
 
 @Module({
   imports: [
+    AuthModule,
     CommentsModule,
     DataloaderModule,
     EventsModule,
